clean(Wallet): Moved checkWalletChecksum to WalletNetwork
