+++ src/wallet-network.js
@@ -82,6 +82,18 @@ function recoverGuid (user_email, captcha) {
     .then(handleResponse).catch(handleError('Could not send recovery email'));
 }
 
+function checkWalletChecksum (payload_checksum, success, error) {
+  var data = {method : 'wallet.aes.json', format : 'json', checksum : payload_checksum};
+
+  API.securePostCallbacks('wallet', data, function (obj) {
+    if (!obj.payload || obj.payload == 'Not modified') {
+      if (success) success();
+    } else if (error) error();
+  }, function () {
+    if (error) error();
+  });
+}
+
 /**
  * Trigger the 2FA reset process
  * @param {string} user_guid User GUID.
@@ -169,6 +181,7 @@ function insertWallet (guid, sharedKey, password, extra, decryptWalletProgress) 
 }
 
 module.exports = {
+  checkWalletChecksum: checkWalletChecksum,
   insertWallet: insertWallet,
   generateUUIDs: generateUUIDs,
   resendTwoFactorSms: resendTwoFactorSms,
